Start turning static lib into dll

+++ JF2/src/JF2/resource_manager.cpp
@@ -33,7 +33,7 @@ const std::shared_ptr<TextRenderer>& ResourceManager::LoadTextRenderer(
 	const std::shared_ptr<Shader>& shader,
 	size_t width,
 	size_t height,
-	size_t pixel_size,
+	unsigned int pixel_size,
 	const std::string& file_name,
 	std::string name) {
